+++ changelog.md
@@ -4,6 +4,7 @@
 * Added a message when the first dragon is killed or the first wither is spawned mentioning that the next ones will be harder to fight
 * Added a config option to Ignore Witherproof Blocks (disabled by default)  
   When enabled the wither will destroy wither proof blocks too (still will not destroy unbreakable blocks)
+* Wither no longer dies when starting a charge attack if he has less than 10 health
 * Fixed crash when min cooldown was higher than max cooldown
 
 ## 3.3.2