+++ changelog.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## Upcoming
+## 3.3.3
 * Wither will now try to charge the target when hidden for too long
-* Added a message when the first dragon is killed or the first wither is spawned mentioning that the next ones will be harder to fight
+* Added a message when the first dragon is killed or the first wither is summoned mentioning that the next ones will be harder to fight
 * Added a config option to Ignore Witherproof Blocks (disabled by default)  
   When enabled the wither will destroy wither proof blocks too (still will not destroy unbreakable blocks)
 * Wither no longer dies when starting a charge attack if he has less than 10 health